Refactor: Move converter classes to converters package.
